Revert "Feature/Jwt prototype"
