+++ src/main/java/com/maze/back/maps/MainMap.java
@@ -34,7 +34,7 @@ public class MainMap {
 		else
 			res.put("up", false);
 		
-		if(x<map.length) {
+		if(x<7) {
 			if(map[x+1][y]==0) {
 				res.put("down", true);
 			}
@@ -54,7 +54,7 @@ public class MainMap {
 		else
 			res.put("left", false);
 		
-		if(y<map[0].length) {
+		if(y<7) {
 			if(map[x][y+1]==0) {
 				res.put("right", true);
 			}
