--- src/main/java/com/maze/back/security/JwtAuthenticationFilter.java
@@ -1,58 +1,0 @@
-package com.maze.back.security;
-
-import java.io.IOException;
-import java.sql.Date;
-import java.util.ArrayList;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
-	private final AuthenticationManager authenticationManager;
-	
-	@Override
-	public Authentication attemptAuthentication(HttpServletRequest request,HttpServletResponse response) throws AuthenticationException{
-		LoginViewModel credentials=null;
-		
-		try {
-			credentials=new ObjectMapper().readValue(request.getInputStream(),LoginViewModel.class);
-		}catch(IOException e) {
-			e.printStackTrace();
-		}
-		
-		UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
-				credentials.getUsername(),
-				credentials.getPassword(),
-				new ArrayList<>()
-				);
-		Authentication auth=authenticationManager.authenticate(authenticationToken);
-		return auth;
-		
-	}
-	@Override
-	protected void successfulAuthentication(HttpServletRequest request,HttpServletResponse response,FilterChain chain, Authentication authResult) throws IOException, ServletException{
-		UserPrincipal principal = (UserPrincipal) authResult.getPrincipal();
-		
-		String token=JWT.create()
-				.withSubject(principal.getUsername())
-				.withExpiresAt(new Date(System.currentTimeMillis()+JwtProperties.EXPIRATION_TIME))
-				.sign(Algorithm.HMAC512(JwtProperties.SECRET.getBytes()));
-		
-		response.addHeader(JwtProperties.HEADER_STRING, JwtProperties.TOKEN_PREFIX+token);
-	}
-}
