--- src/main/java/com/maze/back/security/JwtProperties.java
@@ -1,8 +1,0 @@
-package com.maze.back.security;
-
-public class JwtProperties {
-	public static final String SECRET="root";
-	public static final int EXPIRATION_TIME=300000;
-	public static final String TOKEN_PREFIX="Ugongjae";
-	public static final String HEADER_STRING="Authorization";
-}
