--- src/main/java/com/maze/back/security/LoginViewModel.java
@@ -1,11 +1,0 @@
-package com.maze.back.security;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginViewModel {
-	private String username;
-	private String password;
-}
