--- src/main/java/com/maze/back/security/UserPrincipalDetailsService.java
@@ -1,24 +1,0 @@
-package com.maze.back.security;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserPrincipalDetailsService implements UserDetailsService {
-
-    private UserRepository userRepository;
-
-    public UserPrincipalDetailsService(UserRepository userRepository){
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = this.userRepository.findByUsername(username);
-        UserPrincipal userPrincipal = new UserPrincipal(user);
-
-        return userPrincipal;
-    }
-}
