+++ src/main/resources/application.properties
@@ -1,5 +1,1 @@
 
-spring.datasource.driver-class-name=com.mysql.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3306/maze?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=1234
