--- src/main/resources/mybatis/mapper/UserMapper.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.simplify.sample.security.UserMapper">
-    <select id="readUser" parameterType="String" resultType="com.simplify.sample.security.Member">
-        SELECT * FROM user WHERE id = #{id}
-    </select>
-</mapper>
