--- src/main/resources/templates/adminpage.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Admin Home Page</title>
-</head>
-<body>
-<h1>Admin Page</h1>
-</body>
-</html>
