--- src/main/resources/templates/memberJoinForm.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-    <meta charset="UTF-8">
-    <title>Join</title>
-</head>
-<body>
-
-<form th:object="${member}" th:action="@{/member/new}" th:method="post">
-    <div><label>아이디
-        <input type="text" th:field="*{id}">
-    </label></div>
-    <div><label>비밀번호
-        <input type="text" th:field="*{pw}">
-    </label></div>
-    <button type="submit">저장</button>
-</form>
-
-</body>
-</html>
